Add ability to generate changes file when using svn
Add detect_changes_commands_svn, get_svn_log, get_svn_rev functions.
Move changesgenerate and servicedata tests from gittests.py to gitsvntest.py
so both git and svn canuse these tests.
Update svntests.py and gittests.py to include new gitsvntests.py.
