+++ tests/gittests.py
@@ -7,11 +7,12 @@ import tarfile
 import textwrap
 
 from githgtests  import GitHgTests
+from gitsvntests import GitSvnTests
 from gitfixtures import GitFixtures
 from utils       import run_git
 
 
-class GitTests(GitHgTests):
+class GitTests(GitHgTests, GitSvnTests):
 
     """Unit tests for 'tar_scm --scm git'.
 
@@ -63,6 +64,21 @@ class GitTests(GitHgTests):
 
     def abbrev_sha1s(self, rev):
         return self.sha1s(rev)[0:7]
+
+    def changesrevision(self, rev, abbrev=False):
+        if abbrev:
+            return self.abbrev_sha1s('tag%d' % rev)
+        return self.sha1s('tag%d' % rev)
+
+    def changesregex(self, rev):
+        return '\d{10}.%s' % rev
+
+    def tar_scm_args(self):
+        scm_args = [
+            '--changesgenerate', 'enable',
+            '--versionformat', '0.6.%h',
+        ]
+        return scm_args
 
     # N.B. --versionformat gets tested thoroughly in githgtests.py
 
@@ -180,196 +196,6 @@ class GitTests(GitHgTests):
         sha1 = m.group(1)
         self.assertEqual(sha1, expected_sha1)
 
-    def test_changesgenerate_disabled(self):
-        self.tar_scm_std('--changesgenerate', 'disable')
-
-    def test_changesgenerate_no_servicedata(self):
-        self.tar_scm_std('--changesgenerate', 'enable')
-        self._check_servicedata()
-
-    def test_changesgenerate_corrupt_servicedata(self):
-        with open(os.path.join(self.pkgdir, '_servicedata'), 'w') as sd:
-            sd.write('this is not valid xml')
-        self.tar_scm_std('--changesgenerate', 'enable', should_succeed=False)
-
-    def test_changesgenerate_empty_servicedata_file(self):
-        sd = open(os.path.join(self.pkgdir, '_servicedata'), 'w')
-        sd.close()
-        self.tar_scm_std('--changesgenerate', 'enable')
-        self._check_servicedata()
-
-    def test_changesgenerate_empty_servicedata_element(self):
-        with open(os.path.join(self.pkgdir, '_servicedata'), 'w') as sd:
-            sd.write("<servicedata>\n</servicedata>\n")
-        self.tar_scm_std('--changesgenerate', 'enable')
-        self._check_servicedata()
-
-    def test_changesgenerate_no_changesrevision(self):
-        with open(os.path.join(self.pkgdir, '_servicedata'), 'w') as sd:
-            sd.write(textwrap.dedent("""\
-              <servicedata>
-                <service name="tar_scm">
-                  <param name="url">%s</param>
-                </service>
-              </servicedata>
-            """ % self.fixtures.repo_url))
-        self.tar_scm_std('--changesgenerate', 'enable')
-        self._check_servicedata()
-
-    def _write_servicedata(self, rev):
-        sha1 = self.sha1s('tag%d' % rev)
-        with open(os.path.join(self.pkgdir, '_servicedata'), 'w') as sd:
-            sd.write(textwrap.dedent("""\
-              <servicedata>
-                <service name="tar_scm">
-                  <param name="url">%s</param>
-                  <param name="changesrevision">%s</param>
-                </service>
-              </servicedata>
-            """ % (self.fixtures.repo_url, sha1)))
-
-    def _write_changes_file(self):
-        contents = textwrap.dedent("""\
-          -------------------------------------------------------------------
-          Fri Oct  3 00:17:50 BST 2014 - %s
-
-          - 2
-
-          -------------------------------------------------------------------
-          Thu Sep 18 10:27:14 BST 2014 - %s
-
-          - 1
-        """ % (self.fixtures.user_email, self.fixtures.user_email))
-        with open(os.path.join(self.pkgdir, 'pkg.changes'), 'w') as f:
-            f.write(contents)
-        return contents
-
-    def test_changesgenerate_no_change_or_changes_file(self):
-        self._write_servicedata(2)
-        self.tar_scm_std('--changesgenerate', 'enable')
-        self._check_servicedata()
-
-    def test_changesgenerate_no_change_same_changes_file(self):
-        self._write_servicedata(2)
-        orig_changes = self._write_changes_file()
-        self.tar_scm_std('--changesgenerate', 'enable')
-        self._check_servicedata()
-
-    def test_changesgenerate_new_commit_no_changes_file(self):
-        self._write_servicedata(2)
-        self.fixtures.create_commits(1)
-        self.tar_scm_std('--changesgenerate', 'enable')
-        self._check_servicedata(revision=3)
-
-    def _test_changesgenerate_new_commit_and_changes_file(self, author=None):
-        self._write_servicedata(2)
-        orig_changes = self._write_changes_file()
-        self.fixtures.create_commits(3)
-
-        tar_scm_args = [
-            '--changesgenerate', 'enable',
-            '--versionformat', '0.6.%h',
-        ]
-        if author is not None:
-            tar_scm_args += ['--changesauthor', self.fixtures.user_email]
-        self.tar_scm_std(*tar_scm_args)
-
-        self._check_servicedata(revision=5, expected_dirents=3)
-
-        expected_author = author or 'opensuse-packaging@opensuse.org'
-        expected_changes_regexp = self._new_change_entry_regexp(
-            expected_author,
-            textwrap.dedent("""\
-              - Update to version 0.6.%s:
-                \* 3
-                \* 4
-                \* 5
-              """) % self.abbrev_sha1s('tag5')
-        )
-        self._check_changes(orig_changes, expected_changes_regexp)
-
-    def test_changesgenerate_new_commit_and_changes_file_no_version(self):
-        self._write_servicedata(2)
-        orig_changes = self._write_changes_file()
-        self.fixtures.create_commits(3)
-
-        tar_scm_args = [
-            '--changesgenerate', 'enable',
-            '--version', '',
-            '--changesauthor', self.fixtures.user_email
-        ]
-        self.tar_scm_std(*tar_scm_args)
-
-        self._check_servicedata(revision=5, expected_dirents=3)
-
-        expected_author = self.fixtures.user_email
-        expected_changes_regexp = self._new_change_entry_regexp(
-            expected_author,
-            textwrap.dedent("""\
-              - Update to version \d{10}.%s:
-                \* 3
-                \* 4
-                \* 5
-              """) % self.abbrev_sha1s('tag5')
-        )
-        self._check_changes(orig_changes, expected_changes_regexp)
-
-    def _new_change_entry_regexp(self, author, changes):
-        return textwrap.dedent("""\
-          ^-------------------------------------------------------------------
-          \w{3} \w{3} [ \d]\d \d\d:\d\d:\d\d [A-Z]{3} 20\d\d - %s
-
-          %s
-          """) % (author, changes)
-
-    def _check_changes(self, orig_changes, expected_changes_regexp):
-        new_changes_file = os.path.join(self.outdir, 'pkg.changes')
-        self.assertTrue(os.path.exists(new_changes_file))
-        with open(new_changes_file) as f:
-            new_changes = f.read()
-            self.assertNotEqual(orig_changes, new_changes)
-            print new_changes
-            expected_changes_regexp += "(.*)"
-            self.assertRegexpMatches(new_changes, expected_changes_regexp)
-            m = re.match(expected_changes_regexp, new_changes, re.DOTALL)
-            self.assertEqual(m.group(1), orig_changes)
-
-    def test_changesgenerate_new_commit_and_changes_file(self):
-        self._test_changesgenerate_new_commit_and_changes_file(
-            self.fixtures.user_email)
-
-    def test_changesgenerate_new_commit_and_changes_file_default_author(self):
-        self._test_changesgenerate_new_commit_and_changes_file()
-
-    def test_changesgenerate_new_commit_and_changes_file_with_subdir(self):
-        self._write_servicedata(2)
-        orig_changes = self._write_changes_file()
-        self.fixtures.create_commits(3)
-        self.fixtures.create_commits(3, subdir='another_subdir')
-
-        tar_scm_args = [
-            '--changesgenerate', 'enable',
-            '--versionformat', '0.6.%h',
-            '--subdir', 'another_subdir',
-            '--changesauthor', self.fixtures.user_email,
-        ]
-
-        self.tar_scm_std(*tar_scm_args)
-
-        self._check_servicedata(revision=8, expected_dirents=3)
-
-        expected_author = self.fixtures.user_email
-        expected_changes_regexp = self._new_change_entry_regexp(
-            expected_author,
-            textwrap.dedent("""\
-              - Update to version 0.6.%s:
-                \* 6
-                \* 7
-                \* 8
-              """) % self.abbrev_sha1s('tag8')
-        )
-        self._check_changes(orig_changes, expected_changes_regexp)
-
     def test_updatecache_has_tag(self):
         fix = self.fixtures
         fix.create_commits(2)
@@ -383,7 +209,3 @@ class GitTests(GitHgTests):
         fix.safe_run('checkout tag2')
         fix.create_commits(3)
         fix.safe_run('tag -a -m some_message detached_tag')
-
-        self.tar_scm_std("--revision", 'detached_tag',
-                         "--versionformat", "@PARENT_TAG@")
-        self.assertTarOnly(self.basename(version="detached_tag"))
