+++ tests/svntests.py
@@ -1,11 +1,15 @@
 #!/usr/bin/env python2
 
-from commontests import CommonTests
+import os
+import re
+import textwrap
+
+from gitsvntests import GitSvnTests
 from svnfixtures import SvnFixtures
 from utils       import run_svn
 
 
-class SvnTests(CommonTests):
+class SvnTests(GitSvnTests):
 
     """Unit tests for 'tar_scm --scm svn'.
 
@@ -22,6 +26,19 @@ class SvnTests(CommonTests):
     def default_version(self):
         return self.rev(2)
 
+    def changesrevision(self, rev, abbrev=False):
+        return rev
+
+    def changesregex(self, rev):
+        return rev
+
+    def tar_scm_args(self):
+        scm_args = [
+            '--changesgenerate', 'enable',
+            '--versionformat', '0.6.%r',
+        ]
+        return scm_args
+
     def test_versionformat_rev(self):
         self.tar_scm_std('--versionformat', 'myrev%r.svn')
         self.assertTarOnly(self.basename(version='myrev2.svn'))
@@ -36,3 +53,19 @@ class SvnTests(CommonTests):
         basename = self.basename(version='foo2')
         th = self.assertTarOnly(basename)
         self.assertTarMemberContains(th, basename + '/a', '2')
+
+    def _check_servicedata(self, expected_dirents=2, revision=2):
+        dirents = self.assertNumDirents(self.outdir, expected_dirents)
+        self.assertTrue('_servicedata' in dirents,
+                        '_servicedata in %s' % repr(dirents))
+        sd = open(os.path.join(self.outdir, '_servicedata')).read()
+        expected = """\
+          <servicedata>
+            <service name="tar_scm">
+              <param name="url">%s</param>
+              <param name="changesrevision">([0-9].*)</param>
+            </service>
+          </servicedata>""" % self.fixtures.repo_url
+        (expected, count) = re.subn('\s{2,}', '\s*', expected)
+        m = re.match(expected, sd)
+        self.assertTrue(m, "\n'%s'\n!~ /%s/" % (sd, expected))
